refactor: 품목 로딩 UI 컴포넌트 정리
- 품목 검색/상세 로딩 표시를 전용 컴포넌트로 통일

- 품목 상세 로딩 상태에도 접근성 설명 적용

- presentation 컴파일로 변경 범위 확인

Refs #141
