+++ presentation/src/main/java/com/team/yeogibeoryeo/presentation/search/components/ItemSearchStatusContent.kt
@@ -1,12 +1,15 @@
 package com.team.yeogibeoryeo.presentation.search.components
 
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -14,6 +17,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
@@ -80,6 +85,25 @@ fun ItemSearchStatusDescription(
     )
 }
 
+@Composable
+fun ItemSearchLoadingContent(
+    modifier: Modifier = Modifier,
+) {
+    val loadingContentDescription = stringResource(R.string.loading_action)
+
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center,
+    ) {
+        CircularProgressIndicator(
+            modifier = Modifier.semantics {
+                contentDescription = loadingContentDescription
+            },
+            color = MaterialTheme.colorScheme.primary,
+        )
+    }
+}
+
 @Preview(showBackground = true)
 @Composable
 private fun ItemSearchStatusContentEmptyPreview() {
@@ -95,6 +119,16 @@ private fun ItemSearchStatusContentEmptyPreview() {
                     )
                 },
             )
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun ItemSearchLoadingContentPreview() {
+    YeogiBeoryeoTheme {
+        Surface {
+            ItemSearchLoadingContent()
         }
     }
 }
